Increased timeout for item commands and updates to 30 seconds (#4679)
* Increased timeout for item commands and updates to 30 seconds

+++ bundles/core/org.eclipse.smarthome.core/src/main/java/org/eclipse/smarthome/core/common/SafeCallerBuilder.java
@@ -40,7 +40,7 @@ public interface SafeCallerBuilder<T> {
      * @param timeout the timeout in milliseconds.
      * @return the SafeCallerBuilder itself
      */
-    SafeCallerBuilder<T> withTimeout(int timeout);
+    SafeCallerBuilder<T> withTimeout(long timeout);
 
     /**
      * Specifies the identifier for the context in which only one thread may be occupied at the same time.
